+++ backend/api/admin.py
@@ -3,6 +3,6 @@ from .models import *
 admin.site.register(Quiz)
 admin.site.register(Lesson)
 admin.site.register(Course)
-admin.site.register(Module)
+# admin.site.register(Module)
 
 # Register your models here.
