+++ backend/api/templates/reset_password_email.html
@@ -1,4 +1,4 @@
-
+{% load static %}
 <!DOCTYPE html>
 <html>
 <head>
@@ -13,6 +13,10 @@
             color: #333;
             margin: 0;
             padding: 0;
+			height: 100vh;
+			width: 100%;
+
+		
         }
         h1 {
             font-size: 24px;
@@ -23,7 +27,7 @@
             margin: 0 0 20px;
         }
         a {
-            color: #007bff;
+            color: white;
             text-decoration: none;
         }
         .container {
@@ -32,7 +36,10 @@
             margin: 0 auto;
             padding: 20px;
             box-sizing: border-box;
+			background-color: #98efda;
+			
         }
+		a.button,
         .button {
             display: inline-block;
             background-color: #007bff;
@@ -44,6 +51,7 @@
         }
         .button:hover {
             background-color: #0056b3;
+		
         }
     </style>
 </head>
@@ -51,8 +59,8 @@
     <div class="container">
         <h1>Réinitialisation de mot de passe</h1>
         <p>Vous avez demandé la réinitialisation de votre mot de passe. Veuillez cliquer sur le lien ci-dessous pour créer un nouveau mot de passe :</p>
-        <p><a href="{{ reset_password_link }}" class="button">{{ reset_password_link }}</a></p>
-        <p>Ce lien expire dans {{ timeout }} heures.</p>
+        <p><a href="{{ reset_password_link }}" class="button">  Cliquez Ici </a></p>
+        <!-- <p>Ce lien expire dans {{ timeout }} heures.</p> -->
         <p>Si vous n'avez pas demandé la réinitialisation de votre mot de passe, ignorez simplement cet e-mail.</p>
     </div>
 </body>
