--- views.py
@@ -1,209 +1,0 @@
-from rest_framework.decorators import api_view, permission_classes
-from rest_framework.response import Response
-from rest_framework.permissions import IsAuthenticated
-from .models import *
-from .serializers import *
-
-
-@api_view(['GET'])
-@permission_classes([IsAuthenticated])
-def user_list(request):
-    users = User.objects.all()
-    serializer = UserSerializer(users, many=True)
-    return Response(serializer.data)
-
-@api_view(['GET'])
-@permission_classes([IsAuthenticated])
-def category_list(request):
-    categories = Category.objects.all()
-    serializer = CategorySerializer(categories, many=True)
-    return Response(serializer.data)
-
-@api_view(['GET', 'POST'])
-@permission_classes([IsAuthenticated])
-def course_list(request):
-    if request.method == 'GET':
-        courses = Course.objects.all()
-        serializer = CourseSerializer(courses, many=True)
-        return Response(serializer.data)
-    elif request.method == 'POST':
-        serializer = CourseSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=201)
-        return Response(serializer.errors, status=400)
-
-@api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes([IsAuthenticated])
-def course_detail(request, pk):
-    try:
-        course = Course.objects.get(pk=pk)
-    except Course.DoesNotExist:
-        return Response(status=404)
-
-    if request.method == 'GET':
-        serializer = CourseSerializer(course)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = CourseSerializer(course, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=400)
-    elif request.method == 'DELETE':
-        course.delete()
-        return Response(status=204)
-
-@api_view(['GET', 'POST'])
-@permission_classes([IsAuthenticated])
-def lesson_list(request):
-    if request.method == 'GET':
-        lessons = Lesson.objects.all()
-        serializer = LessonSerializer(lessons, many=True)
-        return Response(serializer.data)
-    elif request.method == 'POST':
-        serializer = LessonSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=201)
-        return Response(serializer.errors, status=400)
-
-@api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes([IsAuthenticated])
-def lesson_detail(request, pk):
-    try:
-        lesson = Lesson.objects.get(pk=pk)
-    except Lesson.DoesNotExist:
-        return Response(status=404)
-
-    if request.method == 'GET':
-        serializer = LessonSerializer(lesson)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = LessonSerializer(lesson, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=400)
-    elif request.method == 'DELETE':
-        lesson.delete()
-        return Response(status=204)
-
-@api_view(['GET', 'POST'])
-@permission_classes([IsAuthenticated])
-def quiz_list(request):
-    if request.method == 'GET':
-        quizzes = Quiz.objects.all()
-        serializer = QuizSerializer(quizzes, many=True)
-        return Response(serializer.data)
-    elif request.method == 'POST':
-        serializer = QuizSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=201)
-        return Response(serializer.errors, status=400)
-
-@api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes([IsAuthenticated])
-def quiz_detail(request, pk):
-    try:
-        quiz = Quiz.objects.get(pk=pk)
-    except Quiz.DoesNotExist:
-        return Response(status=404)
-
-    if request.method == 'GET':
-        serializer = QuizSerializer(quiz)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = QuizSerializer(quiz, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=400)
-    elif request.method == 'DELETE':
-        quiz.delete()
-        return Response(status=204)
-
-@api_view(['GET', 'POST'])
-@permission_classes([IsAuthenticated])
-def question_list(request):
-    if request.method == 'GET':
-        questions = Question.objects.all()
-        serializer = QuestionSerializer(questions, many=True)
-        return Response(serializer.data)
-    elif request.method == 'POST':
-        serializer = QuestionSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=201)
-        return Response(serializer.errors, status=400)
-
-@api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes([IsAuthenticated])
-def question_detail(request, pk):
-    try:
-        question = Question.objects.get(pk=pk)
-    except Question.DoesNotExist:
-        return Response(status=404)
-
-    if request.method == 'GET':
-        serializer = QuestionSerializer(question)
-        return Response(serializer.data)
-    elif requestSorry, it seems that the code I was generating got cut off. Here's the full example:
-
-
-@api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes([IsAuthenticated])
-def question_detail(request, pk):
-    try:
-        question = Question.objects.get(pk=pk)
-    except Question.DoesNotExist:
-        return Response(status=404)
-
-    if request.method == 'GET':
-        serializer = QuestionSerializer(question)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = QuestionSerializer(question, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=400)
-    elif request.method == 'DELETE':
-        question.delete()
-        return Response(status=204)
-
-@api_view(['GET', 'POST'])
-@permission_classes([IsAuthenticated])
-def enrollment_list(request):
-    if request.method == 'GET':
-        enrollments = Enrollment.objects.all()
-        serializer = EnrollmentSerializer(enrollments, many=True)
-        return Response(serializer.data)
-    elif request.method == 'POST':
-        serializer = EnrollmentSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=201)
-        return Response(serializer.errors, status=400)
-
-@api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes([IsAuthenticated])
-def enrollment_detail(request, pk):
-    try:
-        enrollment = Enrollment.objects.get(pk=pk)
-    except Enrollment.DoesNotExist:
-        return Response(status=404)
-
-    if request.method == 'GET':
-        serializer = EnrollmentSerializer(enrollment)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = EnrollmentSerializer(enrollment, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=400)
-    elif request.method == 'DELETE':
-        enrollment.delete()
-        return Response(status=204)
